split out code into components, use vector<TripInfo> in FileParser
