+++ src/file_parser.cpp
@@ -14,108 +14,6 @@
 #include "file_parser.hh"
 
 
-struct PositionInfo{
-    std::string current_stop;
-    int timestamp;
-};
-
-struct TripInfo {
-    std::string trip_id;
-    std::string direction;
-    std::string train_line;
-    int hours;
-    int minutes;
-    std::string str_hours_minutes;
-    PositionInfo pi;
-    bool is_assigned;
-};
-
-
-// Copied from the interwebz (how is this not standardized is beyond me)
-std::string get_time(time_t unix_timestamp){
-
-    if (unix_timestamp == -1){
-        return std::string("<not reported>");
-    }
-
-    char time_buf[80];
-    struct tm ts;
-    ts = *localtime(&unix_timestamp);
-    strftime(time_buf, sizeof(time_buf), "%m-%d-%Y %H:%M:%S %Z", &ts);
-    return std::string(time_buf);
-}
-
-std::ostream& operator<<(std::ostream& os, TripInfo& ti){
-    std::stringstream ss;
-    ss << (ti.is_assigned ? "+ " : "- ")
-       << "Train: " << ti.trip_id << ", " << ti.direction << " " << ti.train_line 
-       << " train, departed at: " 
-       << std::setfill('0') << std::setw(2) << ti.hours << ":" 
-       << std::setfill('0') << std::setw(2) << ti.minutes
-       << ". Current position: " << ti.pi.current_stop
-       << " @ " << get_time(ti.pi.timestamp);
-    os << ss.str();
-    return os;
-}
-
-bool sortDepartureTime(const TripInfo& a, const TripInfo& b){
-    if (a.hours == b.hours){
-        return a.minutes < b.minutes;
-    }
-    return a.hours < b.hours;
-}
-
-bool get_TripInfo(std::string str, TripInfo& ti){
-    
-    // Parses the raw trip ID into something human-readable
-    // Per https://api.mta.info/GTFS.pdf
-    // Initial 6 digits indicate origin time
-        // Times are coded reflecting hundredths of a minute past midnight
-        // '+' indicates 30 seconds
-        // Rounding to nearest minute is acceptable as this is more precise than a transit operation can actually operate
-        // Negative numbers reflect times prior to the day of the schedule
-        // Numbers in excess of '00144000' (1440 minutes per day) reflect times beyond the day of schedule
-        // 00145000 indicates 12:10AM tomorrow
-    // Underscore is a separator
-    // "2..N08R" identifies trip path (stopping pattern) for a unique train trip
-        // Decomposing: 2 -> 2 train; N -> Direction (North), 08R -> Path identifier
-        // Path identifier encodes origin, destination, stops, express/local, operating time periods, and shape (circle / diamond).
-        // Path identifier is OPTIONAL is may not always be present, it may be present in some messages and not present in others
-
-    // ******* NOTE *******
-    // There is almost certainly a library that does this 
-    // Need to replace this in the future, as this is extremly brittle
-    // ******* NOTE *******
-
-    // Onto the parsing per above rules
-    // Just assign the raw trip_id first
-    ti.trip_id = str;
-    
-    // origin departure time info
-    // remember that it's stored as HUNDREDTHS of minutes past midnight, so convert
-    float minutes_past_midnight = atoi(str.substr(0,6).c_str()) / 100;
-    float hours_past_midnight = floor(minutes_past_midnight / 60);
-    float minutes_past_hour = minutes_past_midnight - (60*hours_past_midnight);
-    std::stringstream hours_and_minutes;
-    hours_and_minutes << std::setfill('0') << std::setw(2) << hours_past_midnight;
-    hours_and_minutes << std::setfill('0') << std::setw(2) << minutes_past_hour;
-
-    // Train line
-    std::string train_line(str.substr(7,1));
-
-    // Direction
-    std::string train_direction(str.substr(str.size()-1,1));
-    std::string dir = train_direction == std::string("N") ? "Northbound" : "Southbound";
-
-    ti.direction = dir;
-    ti.train_line = train_line;
-    ti.hours = hours_past_midnight;
-    ti.minutes = minutes_past_hour;
-    ti.str_hours_minutes = hours_and_minutes.str();
-    
-    return true;
-}
-
 // Parse file into protobuf objects
 bool FileParser::parse_file(StaticData* sd)
 {
@@ -132,7 +30,6 @@ bool FileParser::parse_file(StaticData* sd)
         }
         
         transit_realtime::FeedMessage fmessage;
-        std::vector<TripInfo> all_trips;
         if (!fmessage.ParseFromIstream(&data))
         {
             std::cerr << "Error parsing file: " << k_filepath << ", error: " << strerror(errno) << std::endl;
@@ -181,7 +78,7 @@ bool FileParser::parse_file(StaticData* sd)
                         ti.pi.current_stop = sd->get_stop_name(vehicle_pos.stop_id());
                         ti.pi.timestamp = vehicle_pos.has_timestamp() ? vehicle_pos.timestamp() : -1;
                     }
-                    all_trips.push_back(ti);
+                    k_all_trips.push_back(ti);
                 }
                 
 
@@ -191,10 +88,10 @@ bool FileParser::parse_file(StaticData* sd)
 
         // sort the trips based on departure time
         // then iter and dump them out
-        std::sort(all_trips.begin(), all_trips.end(), sortDepartureTime);
+        std::sort(k_all_trips.begin(), k_all_trips.end(), sortDepartureTime);
 
-        for (int i = 0; i < all_trips.size(); i++){
-            std::cout << all_trips[i] << std::endl;
+        for (int i = 0; i < k_all_trips.size(); i++){
+            std::cout << k_all_trips[i] << std::endl;
         }
         
         // TODO: decided what to do with the below, presently unused and might not need to be
@@ -204,7 +101,7 @@ bool FileParser::parse_file(StaticData* sd)
         //std::string json;
         //google::protobuf::util::MessageToJsonString(trip_descriptions, &json);
         //std::cout << json << std::endl;
-        std::cout << "Num trips in this file: " << all_trips.size() << std::endl;
+        std::cout << "Num trips in this file: " << k_all_trips.size() << std::endl;
     }
     else 
     {
